feat: simplify provider
